requirements_mac.txt
toml
lz4
hexdump
tqdm
colorama
InquirerPy
frida
applescript
capstone
keystone-engine